KmiDi_BACKUP/project/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kmidi"
version = "1.0.0"
description = "KmiDi: Unified Music Intelligence & Audio Workstation"
readme = "KMIDI_README.md"
license = "MIT"
authors = [
    {name = "Kelly Project Team"},
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "torch>=2.0",
    "librosa>=0.10",
    "pyyaml>=6.0",
    "scipy>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    # MCP server dependencies (needed for mcp_penta_swarm module)
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-generativeai>=0.5.0",
    "httpx>=0.25.0",
]
audio = [
    "soundfile>=0.12",
    "pydub>=0.25",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
]
mcp = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-generativeai>=0.5.0",
    "httpx>=0.25.0",
]
all = [
    "kmidi[dev,audio,docs,mcp]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["music_brain*", "mcp_workstation*", "mcp_todo*", "mcp_penta_swarm*", "penta_core*"]

[project.scripts]
mcp-penta-swarm = "mcp_penta_swarm.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
